Homework 4 completed
