+++ react1/src/routes/NotFound/index.js
@@ -1,0 +1,7 @@
+export const NotFound = () => {
+    return (
+        <div className="not-found">
+            <h1>404 NOT FOUND</h1>
+        </div>
+    );
+}
