+++ react1/src/routes/index.js
@@ -1,0 +1,4 @@
+export * from './Home';
+export * from './Chats';
+export * from './Profile';
+export * from './NotFound';
